api_yamdb/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.1
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.3
iniconfig==1.1.1
gunicorn==20.0.4
flake8==4.0.1
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
toml==0.10.2
tzdata==2022.1
sqlparse==0.3.1
urllib3==1.26.11
